% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmEM.R
\name{tagmEM}
\alias{tagmEM}
\title{Function to perform generalised t-augmented Gaussian mixture modelling using expectation-maximisation}
\usage{
tagmEM(
  data,
  obs_names = NULL,
  max_iter = 5000,
  tol = 1e-05,
  cluster_sizes = NULL,
  sigk_thresh = 1e-05,
  junk_mixture = TRUE,
  df = 4,
  junk_mean = NULL,
  junk_sd = NULL,
  stop_bic_iter = 5,
  min_clust_search = 8,
  results_list = list("all", "best"),
  rand_sample = seq(0.05, 0.4, by = 0.05)
)
}
\arguments{
\item{data}{numeric object of univariate observations}

\item{obs_names}{character vector of length the number of observations}

\item{max_iter}{numeric integer denoting the maximum number of iterations
before stopping the EM-algorithm's search for a maxima in the
log-likelihood.}

\item{tol}{numeric scalar denoting the maximum absolute difference between
two computations of the log-likelihood with which we accept that a maxima in
the log-likelihood has been computed.}

\item{cluster_sizes}{integer varying from 0 to number of data points N}

\item{sigk_thresh}{lower bound of estimated cluster standard deviation. 
NOTE: avoids singular estimates of the 'sd'}

\item{junk_mixture}{default is TRUE}

\item{df}{default is 'df = 4'}

\item{junk_mean}{numeric scalar denoting the mean of the generalised
t-distribution. By default mean is set to zero.}

\item{junk_sd}{numeric scalar denoting the scale parameter in the generalised
t-distribution}

\item{stop_bic_iter}{numeric integer I, for computational efficiency -
particularly when analysing large numbers of variants - we can stop the
EM-algorithm if the BIC is monotonic increasing over the previous I increases
 in the number of clusters K. By default evidence supporting at least 10
 clusters in the data is computed and so, for example, if the BIC from models
  which assume 6 clusters; 7 clusters; ... or; 10 clusters is monotonic
  increasing - in the number of clusters K -then the EM-algorithm is stopped
  and the model whose K minimises the BIC is returned.}

\item{min_clust_search}{numeric integer which denotes the minimum number of
clusters searched for in the data - default computes evidence supporting up
to K=10 clusters which might explain any clustered heterogeneity in the data.}

\item{results_list}{character list allowing users to choose whether to return
a table with the variants assigned to: "all" of the clusters; a single
"best" cluster or; both. By default we return both, i.e. results_list =
list("all", "best").}

\item{rand_sample}{random probability of being assignment to junk cluster}
}
\value{
Returned are: estimates of the putative number of clusters in the
sample, allocation probabilities and summaries of the
association estimates for each observation;
}
\description{
Function to perform generalised t-augmented Gaussian mixture modelling using expectation-maximisation
}
